+++ Broker_Ammo.lua
@@ -1,11 +1,11 @@
 local frame = CreateFrame("Frame")
-local dataObject = LibStub("LibDataBroker-1.1"):NewDataObject("Broker_Ammo", {type = "data source", text = ""})
 
 local CTimerAfter = C_Timer.After
 local GetInventoryItemCount = GetInventoryItemCount
 local GetInventoryItemTexture = GetInventoryItemTexture
 
 local ammoSlotId
+local dataObject
 
 local function CheckAmmo()
 	if (ammoSlotId ~= nil) then
@@ -26,8 +26,23 @@ local function CheckAmmo()
 end
 
 local function Initialize()
-	ammoSlotId = GetInventorySlotInfo("AMMOSLOT")
-	CheckAmmo()
+	local ammoClasses = {
+		1, -- Warrior
+		3, -- Hunter
+		4, -- Rogue
+	}
+
+	local _, _, classId = UnitClass("player")
+
+	if (tContains(ammoClasses, classId)) then
+		dataObject = LibStub("LibDataBroker-1.1"):NewDataObject("Broker_Ammo", {type = "data source", text = ""})
+		ammoSlotId = GetInventorySlotInfo("AMMOSLOT")
+
+		frame:RegisterEvent("UNIT_INVENTORY_CHANGED")
+		frame:RegisterEvent("BAG_UPDATE")
+
+		CheckAmmo()
+	end
 end
 
 local eventHandlers = {
@@ -36,5 +51,5 @@ local eventHandlers = {
 	BAG_UPDATE = CheckAmmo
   }
 
-  for event in pairs(eventHandlers) do frame:RegisterEvent(event) end
-  frame:SetScript("OnEvent", function (_, event, ...) eventHandlers[event](...) end)
+frame:RegisterEvent("PLAYER_LOGIN")
+frame:SetScript("OnEvent", function (_, event, ...) eventHandlers[event](...) end)