+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# v3
+- Load only for classes that use range ammo
+
 # v2
 - Use event-based approach instead of polling